Added to enable ability to embed Helvetica font into PDF
As Helvetica is a default font of PDFs, it by right cannot be embedded
into the PDF, but rather as code (which many PDF readers would interpret
to output Helvetica). However, as we want to ensure Helvetica is
viewable even on poor quality PDF readers like that on phones, we will
add our own Helvetic .ttf and .otf files, which the Java script can use
to embed Helvetica into the PDF files directly. This is also useful as
our own Helvetica .ttf and .otf files can deal with a wider range of
characters (e.g. German words - see stackoverflow article
https://stackoverflow.com/questions/66966573/flyingsaucerpdf-greek-alphabet-helvetica)