IMPT UPDATE:
1. Added ability to embed Helvetica into the PDF
2. Fixed issue where external css was not read because a DASH "-" is in
the .css file name!
3. Code tested to be working, final product is a 7ed Mastering
Psychiatry PDF that looks exactly like that on the website, but with
better fonts and A4-sized layout (with 1 cm margins)
4. Added code to ensure that if accessing a page on the NUS Medicine's
website causes an error, the script will automatically refresh and try
again. The accept alert function does not seem to work, but it does not
matter as the refresh function always works.
5. Added Regex to handle:
input: "radio" in all the end-of-topic quizzes -> changed to "[___]"
(rudimentary boxes), as input: "radio" always disappeared in the ouput
PDF. (whereas input: "checkbox" always works for some reason). Seems to
be a problem with iText5.
6. Added TimeUnit.SECONDS.sleep(2); to prevent the script from failing -
to prevent the script from immediately taking the HTML code of the
previous page just as the website is loading the next HTML page.
7. Created get_external_css_filepath - so the script automaically
chooses the right external .css file depending on which pages are being
converted into PDF to match the styles as that on the website